dac has to be used in 32bit mode

+++ src/globals.c
@@ -7,7 +7,7 @@ QueueHandle_t queue = NULL;
 TaskHandle_t task_read_handle = NULL;
 TaskHandle_t task_write_handle = NULL;
 
-int16_t samples[CHUNK_SIZE];
+int32_t samples[CHUNK_SIZE];
 
 float pot1 = 0;
 float pot2 = 0;
