+++ src/globals.h
@@ -14,7 +14,7 @@ extern QueueHandle_t queue;
 extern TaskHandle_t task_read_handle;
 extern TaskHandle_t task_write_handle;
 
-extern int16_t samples[CHUNK_SIZE];
+extern int32_t samples[CHUNK_SIZE];
 
 extern float pot1;
 extern float pot2;
